cli: split command dispatch paths
