+++ pkg/cli/cli.go
@@ -9,7 +9,6 @@ import (
 	"log/slog"
 	"net"
 	"os"
-	"os/exec"
 	"os/signal"
 	"path/filepath"
 	"sort"
@@ -26,7 +25,6 @@ import (
 	"github.com/psaab/bpfrx/pkg/config"
 	"github.com/psaab/bpfrx/pkg/configstore"
 	"github.com/psaab/bpfrx/pkg/dataplane"
-	dpuserspace "github.com/psaab/bpfrx/pkg/dataplane/userspace"
 	"github.com/psaab/bpfrx/pkg/dhcp"
 	"github.com/psaab/bpfrx/pkg/dhcprelay"
 	"github.com/psaab/bpfrx/pkg/feeds"
@@ -38,7 +36,6 @@ import (
 	"github.com/psaab/bpfrx/pkg/routing"
 	"github.com/psaab/bpfrx/pkg/rpm"
 	"github.com/psaab/bpfrx/pkg/vrrp"
-	"github.com/vishvananda/netlink"
 	"golang.org/x/sys/unix"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/credentials/insecure"
@@ -630,420 +627,6 @@ func (c *CLI) Run() error {
 		c.refreshPrompt()
 	}
 	return nil
-}
-
-var errExit = fmt.Errorf("exit")
-
-func (c *CLI) dispatch(line string) error {
-	// Extract pipe filter (| match, | except, | find, | count, | last, | no-more).
-	// Skip | display set and | compare (handled separately).
-	if cmd, pipeType, pipeArg, ok := extractPipe(line); ok {
-		return c.dispatchWithPipe(cmd, pipeType, pipeArg)
-	}
-
-	if c.store.InConfigMode() {
-		return c.dispatchConfig(line)
-	}
-
-	// For show commands, auto-page output when it exceeds terminal height.
-	if strings.HasPrefix(strings.TrimSpace(line), "show ") {
-		return c.dispatchWithPager(line)
-	}
-
-	return c.dispatchOperational(line)
-}
-
-// extractPipe splits a line at the last "| <filter>" expression.
-// Recognized filters: match, except, find, count, last, no-more.
-// Returns the command part, pipe type, pipe argument, and whether a pipe was found.
-// Skips "| display set" and "| compare" which are handled separately.
-func extractPipe(line string) (string, string, string, bool) {
-	idx := strings.LastIndex(line, " | ")
-	if idx < 0 {
-		return line, "", "", false
-	}
-	cmd := strings.TrimSpace(line[:idx])
-	pipe := strings.TrimSpace(line[idx+3:])
-	parts := strings.SplitN(pipe, " ", 2)
-	pipeType := parts[0]
-	var pipeArg string
-	if len(parts) > 1 {
-		pipeArg = parts[1]
-	}
-
-	switch pipeType {
-	case "match", "grep", "except", "find", "count", "last", "no-more":
-		return cmd, pipeType, pipeArg, true
-	default:
-		// Not a recognized pipe filter (e.g. "| display set", "| compare")
-		return line, "", "", false
-	}
-}
-
-// dispatchWithPipe runs the command and applies the pipe filter to the output.
-func (c *CLI) dispatchWithPipe(cmd, pipeType, pipeArg string) error {
-	// Capture stdout.
-	origStdout := os.Stdout
-	r, w, err := os.Pipe()
-	if err != nil {
-		return fmt.Errorf("pipe: %w", err)
-	}
-	os.Stdout = w
-
-	// Run the inner command.
-	var cmdErr error
-	done := make(chan struct{})
-	go func() {
-		defer close(done)
-		cmdErr = c.dispatch(cmd)
-	}()
-	<-done
-	w.Close()
-	os.Stdout = origStdout
-
-	// Read captured output.
-	output, _ := io.ReadAll(r)
-	r.Close()
-
-	lines := strings.Split(string(output), "\n")
-	// Remove trailing empty line from split
-	if len(lines) > 0 && lines[len(lines)-1] == "" {
-		lines = lines[:len(lines)-1]
-	}
-
-	switch pipeType {
-	case "match", "grep":
-		for _, line := range lines {
-			if strings.Contains(line, pipeArg) {
-				fmt.Fprintln(origStdout, line)
-			}
-		}
-	case "except":
-		for _, line := range lines {
-			if !strings.Contains(line, pipeArg) {
-				fmt.Fprintln(origStdout, line)
-			}
-		}
-	case "find":
-		found := false
-		for _, line := range lines {
-			if !found && strings.Contains(line, pipeArg) {
-				found = true
-			}
-			if found {
-				fmt.Fprintln(origStdout, line)
-			}
-		}
-	case "count":
-		fmt.Fprintf(origStdout, "Count: %d lines\n", len(lines))
-	case "last":
-		n := 10
-		if pipeArg != "" {
-			if v, err := strconv.Atoi(pipeArg); err == nil && v > 0 {
-				n = v
-			}
-		}
-		start := len(lines) - n
-		if start < 0 {
-			start = 0
-		}
-		for _, line := range lines[start:] {
-			fmt.Fprintln(origStdout, line)
-		}
-	case "no-more":
-		// Output without paging (we don't page by default, so just pass through)
-		for _, line := range lines {
-			fmt.Fprintln(origStdout, line)
-		}
-	}
-
-	return cmdErr
-}
-
-// dispatchWithPager runs a show command and pages the output if it exceeds
-// the terminal height. Press space for next page, enter for next line, q to quit.
-func (c *CLI) dispatchWithPager(line string) error {
-	origStdout := os.Stdout
-	r, w, err := os.Pipe()
-	if err != nil {
-		return c.dispatchOperational(line)
-	}
-	os.Stdout = w
-
-	var cmdErr error
-	done := make(chan struct{})
-	go func() {
-		defer close(done)
-		cmdErr = c.dispatchOperational(line)
-	}()
-	<-done
-	w.Close()
-	os.Stdout = origStdout
-
-	output, _ := io.ReadAll(r)
-	r.Close()
-
-	lines := strings.Split(string(output), "\n")
-	// Remove trailing empty line from split
-	if len(lines) > 0 && lines[len(lines)-1] == "" {
-		lines = lines[:len(lines)-1]
-	}
-
-	// Get terminal height
-	termHeight := 24
-	ws, err := unix.IoctlGetWinsize(int(origStdout.Fd()), unix.TIOCGWINSZ)
-	if err == nil && ws.Row > 0 {
-		termHeight = int(ws.Row)
-	}
-	pageSize := termHeight - 1 // leave one line for the --More-- prompt
-
-	// If output fits on screen, just print it
-	if len(lines) <= pageSize {
-		for _, line := range lines {
-			fmt.Fprintln(origStdout, line)
-		}
-		return cmdErr
-	}
-
-	// Page output
-	lineIdx := 0
-	for lineIdx < len(lines) {
-		// Print a page
-		end := lineIdx + pageSize
-		if end > len(lines) {
-			end = len(lines)
-		}
-		for _, line := range lines[lineIdx:end] {
-			fmt.Fprintln(origStdout, line)
-		}
-		lineIdx = end
-
-		if lineIdx >= len(lines) {
-			break
-		}
-
-		// Show --More-- prompt and wait for input
-		fmt.Fprint(origStdout, "\033[7m--More--\033[0m") // inverse video
-		buf := make([]byte, 1)
-		os.Stdin.Read(buf)
-		fmt.Fprint(origStdout, "\r        \r") // clear --More--
-
-		switch buf[0] {
-		case 'q', 'Q':
-			return cmdErr
-		case '\n', '\r':
-			// Show one more line
-			if lineIdx < len(lines) {
-				fmt.Fprintln(origStdout, lines[lineIdx])
-				lineIdx++
-			}
-			// Don't advance by pageSize, just continue the loop which will
-			// show the prompt again immediately
-			continue
-		default:
-			// space or any other key: show next page
-		}
-	}
-	return cmdErr
-}
-
-var operationalCommands = []string{
-	"configure", "show", "clear", "ping", "test", "traceroute",
-	"monitor", "request", "quit", "exit",
-}
-
-func (c *CLI) dispatchOperational(line string) error {
-	parts := strings.Fields(line)
-	if len(parts) == 0 {
-		return nil
-	}
-
-	if parts[0] == "?" || parts[0] == "help" {
-		c.showOperationalHelp()
-		return nil
-	}
-
-	resolved, err := resolveCommand(parts[0], operationalCommands)
-	if err != nil {
-		return err
-	}
-	parts[0] = resolved
-
-	// RBAC permission check
-	if err := c.checkPermission(parts[0]); err != nil {
-		return err
-	}
-
-	switch parts[0] {
-	case "configure":
-		// Block configure mode on secondary node — config changes must
-		// be made on the primary (RG0 is config authority).
-		if c.cluster != nil && !c.cluster.IsLocalPrimary(0) {
-			return fmt.Errorf("error: node is not primary for RG0, configure on the primary node")
-		}
-		if len(parts) >= 2 && parts[1] == "exclusive" {
-			if err := c.store.EnterConfigureExclusive("cli"); err != nil {
-				return err
-			}
-			c.rl.SetPrompt(c.configPrompt())
-			fmt.Println("Entering configuration mode (exclusive)")
-			fmt.Println("[edit]")
-		} else {
-			if err := c.store.EnterConfigure(); err != nil {
-				return err
-			}
-			c.rl.SetPrompt(c.configPrompt())
-			fmt.Println("Entering configuration mode")
-			fmt.Println("[edit]")
-		}
-		return nil
-
-	case "show":
-		return c.handleShow(parts[1:])
-
-	case "clear":
-		return c.handleClear(parts[1:])
-
-	case "ping":
-		return c.handlePing(parts[1:])
-
-	case "traceroute":
-		return c.handleTraceroute(parts[1:])
-
-	case "monitor":
-		return c.handleMonitor(parts[1:])
-
-	case "request":
-		return c.handleRequest(parts[1:])
-
-	case "test":
-		return c.handleTest(parts[1:])
-
-	case "quit", "exit":
-		return errExit
-
-	default:
-		return fmt.Errorf("unknown command: %s", parts[0])
-	}
-}
-
-func (c *CLI) dispatchConfig(line string) error {
-	parts := strings.Fields(line)
-	if len(parts) == 0 {
-		return nil
-	}
-
-	switch parts[0] {
-	case "edit":
-		if len(parts) < 2 {
-			fmt.Println("edit: missing path")
-			return nil
-		}
-		newPath := append(c.store.GetEditPath(), parts[1:]...)
-		c.store.SetEditPath(newPath)
-		c.rl.SetPrompt(c.configPrompt())
-		fmt.Printf("[edit %s]\n", strings.Join(newPath, " "))
-		return nil
-
-	case "top":
-		c.store.NavigateTop()
-		c.rl.SetPrompt(c.configPrompt())
-		fmt.Println("[edit]")
-		return nil
-
-	case "up":
-		c.store.NavigateUp()
-		c.rl.SetPrompt(c.configPrompt())
-		editPath := c.store.GetEditPath()
-		if len(editPath) > 0 {
-			fmt.Printf("[edit %s]\n", strings.Join(editPath, " "))
-		} else {
-			fmt.Println("[edit]")
-		}
-		return nil
-
-	case "set":
-		if len(parts) < 2 {
-			return fmt.Errorf("set: missing path")
-		}
-		fullPath := append(c.store.GetEditPath(), parts[1:]...)
-		return c.store.SetFromInput(strings.Join(fullPath, " "))
-
-	case "delete":
-		if len(parts) < 2 {
-			return fmt.Errorf("delete: missing path")
-		}
-		fullPath := append(c.store.GetEditPath(), parts[1:]...)
-		return c.store.DeleteFromInput(strings.Join(fullPath, " "))
-
-	case "copy", "rename":
-		return c.handleCopyRename(parts)
-
-	case "insert":
-		return c.handleInsert(parts)
-
-	case "show":
-		return c.handleConfigShow(parts[1:])
-
-	case "commit":
-		return c.handleCommit(parts[1:])
-
-	case "rollback":
-		n := 0
-		if len(parts) >= 2 {
-			fmt.Sscanf(parts[1], "%d", &n)
-		}
-		if err := c.store.Rollback(n); err != nil {
-			return err
-		}
-		fmt.Println("configuration rolled back")
-		return nil
-
-	case "load":
-		return c.handleLoad(parts[1:])
-
-	case "run":
-		if len(parts) < 2 {
-			return fmt.Errorf("run: missing command")
-		}
-		return c.dispatchOperational(strings.Join(parts[1:], " "))
-
-	case "annotate":
-		if len(parts) < 3 {
-			fmt.Println("usage: annotate <path> \"comment\"")
-			return nil
-		}
-		line := strings.Join(parts[1:], " ")
-		quoteIdx := strings.Index(line, "\"")
-		if quoteIdx < 0 {
-			fmt.Println("usage: annotate <path> \"comment\"")
-			return nil
-		}
-		pathStr := strings.TrimSpace(line[:quoteIdx])
-		comment := strings.Trim(line[quoteIdx:], "\"")
-		pathParts := append(c.store.GetEditPath(), strings.Fields(pathStr)...)
-		if err := c.store.Annotate(pathParts, comment); err != nil {
-			return err
-		}
-		fmt.Println("annotation set")
-		return nil
-
-	case "exit", "quit":
-		if c.store.IsDirty() {
-			fmt.Println("warning: uncommitted changes will be discarded")
-		}
-		c.store.ExitConfigure()
-		c.rl.SetPrompt(c.operationalPrompt())
-		fmt.Println("Exiting configuration mode")
-		return nil
-
-	case "?", "help":
-		c.showConfigHelp()
-		return nil
-
-	default:
-		return fmt.Errorf("unknown command: %s (in configuration mode)", parts[0])
-	}
 }
 
 func (c *CLI) handleShow(args []string) error {
@@ -1751,250 +1334,6 @@ func (c *CLI) handleConfigShow(args []string) error {
 	return nil
 }
 
-func (c *CLI) handleCopyRename(parts []string) error {
-	cmd := parts[0]
-	toIdx := -1
-	for i, p := range parts {
-		if p == "to" {
-			toIdx = i
-			break
-		}
-	}
-	if toIdx < 2 || toIdx >= len(parts)-1 {
-		fmt.Printf("usage: %s <src-path> to <dst-path>\n", cmd)
-		return nil
-	}
-	srcPath := parts[1:toIdx]
-	dstPath := parts[toIdx+1:]
-	editPath := c.store.GetEditPath()
-	if len(editPath) > 0 {
-		srcPath = append(append([]string{}, editPath...), srcPath...)
-		dstPath = append(append([]string{}, editPath...), dstPath...)
-	}
-	if cmd == "rename" {
-		return c.store.Rename(srcPath, dstPath)
-	}
-	return c.store.Copy(srcPath, dstPath)
-}
-
-// handleInsert handles:
-//
-//	insert <element-path> before|after <ref-identifier>
-//
-// The ref-identifier (e.g., "policy allow-all") is relative to the same parent
-// as the element. The full reference path is constructed by replacing the
-// element's trailing identifier tokens with the ref-identifier tokens.
-func (c *CLI) handleInsert(parts []string) error {
-	// Find "before" or "after" keyword.
-	kwIdx := -1
-	isBefore := false
-	for i, p := range parts {
-		if p == "before" {
-			kwIdx = i
-			isBefore = true
-			break
-		}
-		if p == "after" {
-			kwIdx = i
-			break
-		}
-	}
-	if kwIdx < 2 || kwIdx >= len(parts)-1 {
-		fmt.Println("usage: insert <element-path> before|after <ref-identifier>")
-		return nil
-	}
-	elemPath := parts[1:kwIdx]
-	refTokens := parts[kwIdx+1:]
-	editPath := c.store.GetEditPath()
-	if len(editPath) > 0 {
-		elemPath = append(append([]string{}, editPath...), elemPath...)
-	}
-	// Construct the full reference path: element's parent path + ref tokens.
-	// The ref tokens replace the element's trailing identifier (same keyword + name).
-	if len(refTokens) > len(elemPath) {
-		fmt.Println("error: reference identifier is longer than element path")
-		return nil
-	}
-	parentPath := elemPath[:len(elemPath)-len(refTokens)]
-	refPath := append(append([]string{}, parentPath...), refTokens...)
-	return c.store.Insert(elemPath, refPath, isBefore)
-}
-
-func (c *CLI) handleLoad(args []string) error {
-	if len(args) < 2 {
-		fmt.Println("load:")
-		fmt.Println("  override terminal    Replace candidate with pasted config")
-		fmt.Println("  merge terminal       Merge pasted config into candidate")
-		fmt.Println("  set terminal         Load set commands from terminal")
-		fmt.Println("  override <file>      Replace candidate with file contents")
-		fmt.Println("  merge <file>         Merge file contents into candidate")
-		return nil
-	}
-
-	mode := args[0] // "override", "merge", or "set"
-	if mode != "override" && mode != "merge" && mode != "set" {
-		return fmt.Errorf("load: unknown mode %q (use 'override', 'merge', or 'set')", mode)
-	}
-
-	source := args[1]
-	if mode == "set" && source != "terminal" {
-		return fmt.Errorf("load set: only 'terminal' source is supported")
-	}
-	var content string
-
-	if source == "terminal" {
-		// Read from terminal until a line containing only a single Ctrl-D marker
-		fmt.Println("[Type or paste configuration, then press Ctrl-D on an empty line]")
-		var lines []string
-		for {
-			line, err := c.rl.Readline()
-			if err != nil {
-				// EOF (Ctrl-D)
-				break
-			}
-			lines = append(lines, line)
-		}
-		content = strings.Join(lines, "\n")
-	} else {
-		// Read from file
-		data, err := os.ReadFile(source)
-		if err != nil {
-			return fmt.Errorf("load: %w", err)
-		}
-		content = string(data)
-	}
-
-	if strings.TrimSpace(content) == "" {
-		return fmt.Errorf("load: empty input")
-	}
-
-	switch mode {
-	case "set":
-		count, err := c.store.LoadSet(content)
-		if err != nil {
-			return fmt.Errorf("load set: %w", err)
-		}
-		fmt.Printf("load set complete: %d commands applied\n", count)
-		return nil
-	default:
-		var err error
-		switch mode {
-		case "override":
-			err = c.store.LoadOverride(content)
-		case "merge":
-			err = c.store.LoadMerge(content)
-		}
-		if err != nil {
-			return fmt.Errorf("load %s: %w", mode, err)
-		}
-		fmt.Printf("load %s complete\n", mode)
-		return nil
-	}
-}
-
-func (c *CLI) handleCommit(args []string) error {
-	if len(args) > 0 && args[0] == "check" {
-		_, err := c.store.CommitCheck()
-		if err != nil {
-			return fmt.Errorf("commit check failed: %w", err)
-		}
-		fmt.Println("configuration check succeeds")
-		return nil
-	}
-
-	if len(args) > 0 && args[0] == "comment" {
-		if len(args) < 2 {
-			return fmt.Errorf("usage: commit comment \"description\"")
-		}
-		desc := strings.Join(args[1:], " ")
-		desc = strings.Trim(desc, "\"'")
-
-		diffSummary := c.store.CommitDiffSummary()
-
-		compiled, err := c.store.CommitWithDescription(desc)
-		if err != nil {
-			return fmt.Errorf("commit failed: %w", err)
-		}
-
-		if c.dp != nil {
-			if err := c.applyToDataplane(compiled); err != nil {
-				fmt.Fprintf(os.Stderr, "warning: dataplane apply failed: %v\n", err)
-			}
-		}
-		c.reloadSyslog(compiled)
-		c.refreshPrompt()
-
-		if diffSummary != "" {
-			fmt.Printf("commit complete: %s\n", diffSummary)
-		} else {
-			fmt.Println("commit complete")
-		}
-		return nil
-	}
-
-	if len(args) > 0 && args[0] == "confirmed" {
-		minutes := 10
-		if len(args) >= 2 {
-			if v, err := strconv.Atoi(args[1]); err == nil && v > 0 {
-				minutes = v
-			}
-		}
-
-		compiled, err := c.store.CommitConfirmed(minutes)
-		if err != nil {
-			return fmt.Errorf("commit confirmed failed: %w", err)
-		}
-
-		// Apply to dataplane
-		if c.dp != nil {
-			if err := c.applyToDataplane(compiled); err != nil {
-				fmt.Fprintf(os.Stderr, "warning: dataplane apply failed: %v\n", err)
-			}
-		}
-		c.reloadSyslog(compiled)
-		c.refreshPrompt()
-
-		fmt.Printf("commit confirmed will be automatically rolled back in %d minutes unless confirmed\n", minutes)
-		return nil
-	}
-
-	// Bare commit: if a confirmed commit is pending, confirm it
-	if c.store.IsConfirmPending() {
-		if err := c.store.ConfirmCommit(); err != nil {
-			return fmt.Errorf("confirm commit: %w", err)
-		}
-		fmt.Println("commit confirmed")
-		return nil
-	}
-
-	// Capture diff summary before commit (active will change)
-	diffSummary := c.store.CommitDiffSummary()
-
-	compiled, err := c.store.Commit()
-	if err != nil {
-		return fmt.Errorf("commit failed: %w", err)
-	}
-
-	// Apply to dataplane
-	if c.dp != nil {
-		if err := c.applyToDataplane(compiled); err != nil {
-			fmt.Fprintf(os.Stderr, "warning: dataplane apply failed: %v\n", err)
-		}
-	}
-
-	// Hot-reload syslog clients
-	c.reloadSyslog(compiled)
-	c.refreshPrompt()
-
-	if diffSummary != "" {
-		fmt.Printf("commit complete: %s\n", diffSummary)
-	} else {
-		fmt.Println("commit complete")
-	}
-	return nil
-}
-
-// refreshPrompt re-reads the system hostname and updates the readline prompt.
 func (c *CLI) refreshPrompt() {
 	if h, err := os.Hostname(); err == nil && h != "" {
 		c.hostname = h
@@ -2510,337 +1849,6 @@ type topTalkerEntry struct {
 	fwdPkts, revPkts                  uint64
 	fwdBytes, revBytes                uint64
 	age                               uint64
-}
-
-func (c *CLI) handleClear(args []string) error {
-	clearTree := operationalTree["clear"].Children
-	showHelp := func() {
-		fmt.Println("clear:")
-		writeCompletionHelp(os.Stdout, treeHelpCandidates(clearTree))
-	}
-	if len(args) < 1 {
-		showHelp()
-		return nil
-	}
-
-	switch args[0] {
-	case "arp":
-		return c.handleClearArp()
-	case "ipv6":
-		return c.handleClearIPv6(args[1:])
-	case "security":
-		return c.handleClearSecurity(args[1:])
-	case "firewall":
-		return c.handleClearFirewall(args[1:])
-	case "dhcp":
-		return c.handleClearDHCP(args[1:])
-	case "interfaces":
-		return c.handleClearInterfaces(args[1:])
-	case "system":
-		return c.handleClearSystem(args[1:])
-	default:
-		showHelp()
-		return nil
-	}
-}
-
-func (c *CLI) handleClearSystem(args []string) error {
-	if len(args) < 1 || args[0] != "config-lock" {
-		cmdtree.PrintTreeHelp("clear system:", operationalTree, "clear", "system")
-		return nil
-	}
-	holder, locked := c.store.ConfigHolder()
-	if !locked {
-		fmt.Println("No configuration lock held")
-		return nil
-	}
-	c.store.ForceExitConfigure()
-	fmt.Printf("Configuration lock cleared (was held by %s)\n", holder)
-	return nil
-}
-
-func (c *CLI) handleClearInterfaces(args []string) error {
-	if len(args) >= 1 && args[0] == "statistics" {
-		fmt.Println("Interface statistics counters noted")
-		fmt.Println("(kernel counters are cumulative and cannot be reset)")
-		return nil
-	}
-	cmdtree.PrintTreeHelp("clear interfaces:", operationalTree, "clear", "interfaces")
-	return nil
-}
-
-func (c *CLI) handleClearArp() error {
-	out, err := exec.Command("ip", "-4", "neigh", "flush", "all").CombinedOutput()
-	if err != nil {
-		return fmt.Errorf("flush ARP: %s", strings.TrimSpace(string(out)))
-	}
-	fmt.Println("ARP cache cleared")
-	return nil
-}
-
-func (c *CLI) handleClearIPv6(args []string) error {
-	if len(args) < 1 || args[0] != "neighbors" {
-		cmdtree.PrintTreeHelp("clear ipv6:", operationalTree, "clear", "ipv6")
-		return nil
-	}
-	out, err := exec.Command("ip", "-6", "neigh", "flush", "all").CombinedOutput()
-	if err != nil {
-		return fmt.Errorf("flush IPv6 neighbors: %s", strings.TrimSpace(string(out)))
-	}
-	fmt.Println("IPv6 neighbor cache cleared")
-	return nil
-}
-
-func (c *CLI) handleClearSecurity(args []string) error {
-	if len(args) < 1 {
-		fmt.Println("clear security:")
-		writeCompletionHelp(os.Stdout, treeHelpCandidates(operationalTree["clear"].Children["security"].Children))
-		return nil
-	}
-
-	switch args[0] {
-	case "nat":
-		if len(args) >= 3 && args[1] == "source" && args[2] == "persistent-nat-table" {
-			return c.clearPersistentNAT()
-		}
-		if len(args) >= 2 && args[1] == "statistics" {
-			if c.dp == nil || !c.dp.IsLoaded() {
-				fmt.Println("dataplane not loaded")
-				return nil
-			}
-			if err := c.dp.ClearNATRuleCounters(); err != nil {
-				return fmt.Errorf("clear NAT counters: %w", err)
-			}
-			fmt.Println("NAT translation statistics cleared")
-			return nil
-		}
-		cmdtree.PrintTreeHelp("clear security nat:", operationalTree, "clear", "security", "nat")
-		return nil
-	case "flow":
-		if len(args) < 2 || args[1] != "session" {
-			return fmt.Errorf("usage: clear security flow session [filters...]")
-		}
-		if c.dp == nil || !c.dp.IsLoaded() {
-			fmt.Println("dataplane not loaded")
-			return nil
-		}
-		f := c.parseSessionFilter(args[2:])
-		if f.hasFilter() {
-			return c.clearFilteredSessions(f)
-		}
-		v4, v6, err := c.dp.ClearAllSessions()
-		if err != nil {
-			return fmt.Errorf("clear sessions: %w", err)
-		}
-		fmt.Printf("%d IPv4 and %d IPv6 session entries cleared\n", v4, v6)
-		c.clearPeerSessions(nil)
-		return nil
-
-	case "policies":
-		if len(args) >= 2 && args[1] == "hit-count" {
-			if c.dp == nil || !c.dp.IsLoaded() {
-				fmt.Println("dataplane not loaded")
-				return nil
-			}
-			if err := c.dp.ClearPolicyCounters(); err != nil {
-				return fmt.Errorf("clear policy counters: %w", err)
-			}
-			fmt.Println("policy hit counters cleared")
-			return nil
-		}
-		return fmt.Errorf("usage: clear security policies hit-count")
-
-	case "counters":
-		if c.dp == nil || !c.dp.IsLoaded() {
-			fmt.Println("dataplane not loaded")
-			return nil
-		}
-		if err := c.dp.ClearAllCounters(); err != nil {
-			return fmt.Errorf("clear counters: %w", err)
-		}
-		fmt.Println("all counters cleared")
-		return nil
-
-	default:
-		cmdtree.PrintTreeHelp("clear security:", operationalTree, "clear", "security")
-		return nil
-	}
-}
-
-// clearFilteredSessions deletes sessions matching the given filter.
-func (c *CLI) clearFilteredSessions(f sessionFilter) error {
-	v4Deleted := 0
-	v6Deleted := 0
-
-	// IPv4: collect matching forward keys and derive reverse keys
-	var v4Keys []dataplane.SessionKey
-	var v4RevKeys []dataplane.SessionKey
-	var snatDNATKeys []dataplane.DNATKey
-	_ = c.dp.IterateSessions(func(key dataplane.SessionKey, val dataplane.SessionValue) bool {
-		if val.IsReverse != 0 {
-			return true // skip reverse entries; they'll be cleaned via forward
-		}
-		if !f.matchesV4(key, val) {
-			return true
-		}
-		v4Keys = append(v4Keys, key)
-		// Reverse key for cleanup
-		v4RevKeys = append(v4RevKeys, dataplane.SessionKey{
-			Protocol: key.Protocol,
-			SrcIP:    key.DstIP,
-			DstIP:    key.SrcIP,
-			SrcPort:  key.DstPort,
-			DstPort:  key.SrcPort,
-		})
-		// Track SNAT DNAT entries for cleanup
-		if val.Flags&dataplane.SessFlagSNAT != 0 &&
-			val.Flags&dataplane.SessFlagStaticNAT == 0 {
-			snatDNATKeys = append(snatDNATKeys, dataplane.DNATKey{
-				Protocol: key.Protocol,
-				DstIP:    val.NATSrcIP,
-				DstPort:  val.NATSrcPort,
-			})
-		}
-		return true
-	})
-
-	for _, key := range v4Keys {
-		if err := c.dp.DeleteSession(key); err == nil {
-			v4Deleted++
-		}
-	}
-	for _, key := range v4RevKeys {
-		c.dp.DeleteSession(key)
-	}
-	for _, dk := range snatDNATKeys {
-		c.dp.DeleteDNATEntry(dk)
-	}
-
-	// IPv6: collect matching forward keys
-	var v6Keys []dataplane.SessionKeyV6
-	var v6RevKeys []dataplane.SessionKeyV6
-	var snatDNATKeysV6 []dataplane.DNATKeyV6
-	_ = c.dp.IterateSessionsV6(func(key dataplane.SessionKeyV6, val dataplane.SessionValueV6) bool {
-		if val.IsReverse != 0 {
-			return true
-		}
-		if !f.matchesV6(key, val) {
-			return true
-		}
-		v6Keys = append(v6Keys, key)
-		v6RevKeys = append(v6RevKeys, dataplane.SessionKeyV6{
-			Protocol: key.Protocol,
-			SrcIP:    key.DstIP,
-			DstIP:    key.SrcIP,
-			SrcPort:  key.DstPort,
-			DstPort:  key.SrcPort,
-		})
-		if val.Flags&dataplane.SessFlagSNAT != 0 &&
-			val.Flags&dataplane.SessFlagStaticNAT == 0 {
-			snatDNATKeysV6 = append(snatDNATKeysV6, dataplane.DNATKeyV6{
-				Protocol: key.Protocol,
-				DstIP:    val.NATSrcIP,
-				DstPort:  val.NATSrcPort,
-			})
-		}
-		return true
-	})
-
-	for _, key := range v6Keys {
-		if err := c.dp.DeleteSessionV6(key); err == nil {
-			v6Deleted++
-		}
-	}
-	for _, key := range v6RevKeys {
-		c.dp.DeleteSessionV6(key)
-	}
-	for _, dk := range snatDNATKeysV6 {
-		c.dp.DeleteDNATEntryV6(dk)
-	}
-
-	fmt.Printf("%d IPv4 and %d IPv6 matching sessions cleared\n", v4Deleted, v6Deleted)
-	c.clearPeerSessions(&f)
-	return nil
-}
-
-// clearPeerSessions forwards a clear request to the cluster peer.
-func (c *CLI) clearPeerSessions(f *sessionFilter) {
-	if c.cluster == nil {
-		return
-	}
-	conn := c.dialPeer()
-	if conn == nil {
-		return
-	}
-	defer conn.Close()
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-	req := &pb.ClearSessionsRequest{}
-	if f != nil {
-		if f.srcNet != nil {
-			req.SourcePrefix = f.srcNet.String()
-		}
-		if f.dstNet != nil {
-			req.DestinationPrefix = f.dstNet.String()
-		}
-		if f.proto != 0 {
-			switch f.proto {
-			case 6:
-				req.Protocol = "tcp"
-			case 17:
-				req.Protocol = "udp"
-			case 1:
-				req.Protocol = "icmp"
-			}
-		}
-		req.SourcePort = uint32(f.srcPort)
-		req.DestinationPort = uint32(f.dstPort)
-		req.Application = f.appName
-	}
-	ctx = metadata.AppendToOutgoingContext(ctx, "x-peer-forwarded", "1")
-	_, _ = pb.NewBpfrxServiceClient(conn).ClearSessions(ctx, req)
-}
-
-func (c *CLI) handleClearFirewall(args []string) error {
-	if len(args) < 1 || args[0] != "all" {
-		cmdtree.PrintTreeHelp("clear firewall:", operationalTree, "clear", "firewall")
-		return nil
-	}
-	if c.dp == nil || !c.dp.IsLoaded() {
-		fmt.Println("Dataplane not loaded")
-		return nil
-	}
-	if err := c.dp.ClearFilterCounters(); err != nil {
-		return fmt.Errorf("clear filter counters: %w", err)
-	}
-	fmt.Println("Firewall filter counters cleared")
-	return nil
-}
-
-func (c *CLI) handleClearDHCP(args []string) error {
-	if len(args) < 1 || args[0] != "client-identifier" {
-		cmdtree.PrintTreeHelp("clear dhcp:", operationalTree, "clear", "dhcp")
-		return nil
-	}
-
-	if c.dhcp == nil {
-		fmt.Println("No DHCP clients running")
-		return nil
-	}
-
-	// Optional interface filter: clear dhcp client-identifier interface <name>
-	if len(args) >= 3 && args[1] == "interface" {
-		ifName := args[2]
-		if err := c.dhcp.ClearDUID(ifName); err != nil {
-			return fmt.Errorf("clear DUID: %w", err)
-		}
-		fmt.Printf("DHCPv6 DUID cleared for %s\n", ifName)
-		return nil
-	}
-
-	c.dhcp.ClearAllDUIDs()
-	fmt.Println("All DHCPv6 DUIDs cleared")
-	return nil
 }
 
 func (c *CLI) handleShowNAT(args []string) error {
@@ -3562,1313 +2570,4 @@ func (c *CLI) handleShowServices(args []string) error {
 	default:
 		return fmt.Errorf("unknown services target: %s", args[0])
 	}
-}
-
-func (c *CLI) handlePing(args []string) error {
-	if len(args) == 0 {
-		fmt.Println("usage: ping <target> [count <N>] [source <IP>] [size <N>] [routing-instance <name>]")
-		return nil
-	}
-
-	target := args[0]
-	count := "5"
-	source := ""
-	size := ""
-	vrfName := ""
-
-	for i := 1; i < len(args)-1; i++ {
-		switch args[i] {
-		case "count":
-			count = args[i+1]
-			i++
-		case "source":
-			source = args[i+1]
-			i++
-		case "size":
-			size = args[i+1]
-			i++
-		case "routing-instance":
-			vrfName = args[i+1]
-			i++
-		}
-	}
-
-	var cmdArgs []string
-	if vrfName != "" {
-		cmdArgs = append(cmdArgs, "ip", "vrf", "exec", "vrf-"+vrfName, "ping")
-	} else {
-		cmdArgs = append(cmdArgs, "ping")
-	}
-
-	cmdArgs = append(cmdArgs, "-c", count)
-	if source != "" {
-		cmdArgs = append(cmdArgs, "-I", source)
-	}
-	if size != "" {
-		cmdArgs = append(cmdArgs, "-s", size)
-	}
-	cmdArgs = append(cmdArgs, target)
-
-	ctx, cancel := context.WithTimeout(context.Background(), 120*time.Second)
-	defer cancel()
-	c.cmdMu.Lock()
-	c.cmdCancel = cancel
-	c.cmdMu.Unlock()
-	defer func() {
-		c.cmdMu.Lock()
-		c.cmdCancel = nil
-		c.cmdMu.Unlock()
-	}()
-
-	cmd := exec.CommandContext(ctx, cmdArgs[0], cmdArgs[1:]...)
-	cmd.Stdout = os.Stdout
-	cmd.Stderr = os.Stderr
-	err := cmd.Run()
-	if ctx.Err() != nil {
-		return nil // cancelled by Ctrl-C or timeout
-	}
-	return err
-}
-
-func (c *CLI) handleTraceroute(args []string) error {
-	if len(args) == 0 {
-		fmt.Println("usage: traceroute <target> [source <IP>] [routing-instance <name>]")
-		return nil
-	}
-
-	target := args[0]
-	source := ""
-	vrfName := ""
-
-	for i := 1; i < len(args)-1; i++ {
-		switch args[i] {
-		case "source":
-			source = args[i+1]
-			i++
-		case "routing-instance":
-			vrfName = args[i+1]
-			i++
-		}
-	}
-
-	var cmdArgs []string
-	if vrfName != "" {
-		cmdArgs = append(cmdArgs, "ip", "vrf", "exec", "vrf-"+vrfName, "traceroute")
-	} else {
-		cmdArgs = append(cmdArgs, "traceroute")
-	}
-
-	if source != "" {
-		cmdArgs = append(cmdArgs, "-s", source)
-	}
-	cmdArgs = append(cmdArgs, target)
-
-	ctx, cancel := context.WithTimeout(context.Background(), 120*time.Second)
-	defer cancel()
-	c.cmdMu.Lock()
-	c.cmdCancel = cancel
-	c.cmdMu.Unlock()
-	defer func() {
-		c.cmdMu.Lock()
-		c.cmdCancel = nil
-		c.cmdMu.Unlock()
-	}()
-
-	cmd := exec.CommandContext(ctx, cmdArgs[0], cmdArgs[1:]...)
-	cmd.Stdout = os.Stdout
-	cmd.Stderr = os.Stderr
-	err := cmd.Run()
-	if ctx.Err() != nil {
-		return nil // cancelled by Ctrl-C or timeout
-	}
-	return err
-}
-
-func (c *CLI) clearPersistentNAT() error {
-	if c.dp == nil || c.dp.GetPersistentNAT() == nil {
-		fmt.Println("Persistent NAT table not available")
-		return nil
-	}
-	count := c.dp.GetPersistentNAT().Len()
-	c.dp.GetPersistentNAT().Clear()
-	fmt.Printf("Cleared %d persistent NAT bindings\n", count)
-	return nil
-}
-
-// showVRRP displays VRRP status.
-func matchPolicyAddr(addrs []string, ip net.IP, cfg *config.Config) bool {
-	if len(addrs) == 0 || ip == nil {
-		return true // no filter = match all
-	}
-	for _, a := range addrs {
-		if a == "any" {
-			return true
-		}
-		if cfg.Security.AddressBook == nil {
-			continue
-		}
-		// Check address entries
-		if addr, ok := cfg.Security.AddressBook.Addresses[a]; ok {
-			_, cidr, err := net.ParseCIDR(addr.Value)
-			if err == nil && cidr.Contains(ip) {
-				return true
-			}
-		}
-		// Check address-set entries (recursive)
-		if matchPolicyAddrSet(a, ip, cfg, 0) {
-			return true
-		}
-	}
-	return false
-}
-
-func matchPolicyAddrSet(setName string, ip net.IP, cfg *config.Config, depth int) bool {
-	if depth > 5 || cfg.Security.AddressBook == nil {
-		return false
-	}
-	as, ok := cfg.Security.AddressBook.AddressSets[setName]
-	if !ok {
-		return false
-	}
-	for _, addrName := range as.Addresses {
-		if addr, ok := cfg.Security.AddressBook.Addresses[addrName]; ok {
-			_, cidr, err := net.ParseCIDR(addr.Value)
-			if err == nil && cidr.Contains(ip) {
-				return true
-			}
-		}
-	}
-	for _, nested := range as.AddressSets {
-		if matchPolicyAddrSet(nested, ip, cfg, depth+1) {
-			return true
-		}
-	}
-	return false
-}
-
-// matchPolicyApp checks if a protocol/port matches a list of application references.
-func matchPolicyApp(apps []string, proto string, dstPort int, cfg *config.Config) bool {
-	if len(apps) == 0 || proto == "" {
-		return true // no filter = match all
-	}
-	for _, a := range apps {
-		if a == "any" {
-			return true
-		}
-		// Check predefined and custom applications
-		if matchSingleApp(a, proto, dstPort, cfg) {
-			return true
-		}
-		// Check application sets
-		if cfg.Applications.ApplicationSets != nil {
-			if as, ok := cfg.Applications.ApplicationSets[a]; ok {
-				for _, appRef := range as.Applications {
-					if matchSingleApp(appRef, proto, dstPort, cfg) {
-						return true
-					}
-				}
-			}
-		}
-	}
-	return false
-}
-
-func matchSingleApp(appName, proto string, dstPort int, cfg *config.Config) bool {
-	if cfg.Applications.Applications == nil {
-		return false
-	}
-	app, ok := cfg.Applications.Applications[appName]
-	if !ok {
-		return false
-	}
-	if app.Protocol != "" && !strings.EqualFold(app.Protocol, proto) {
-		return false
-	}
-	if app.DestinationPort != "" && dstPort > 0 {
-		// Simple port match (handle single port or range)
-		if strings.Contains(app.DestinationPort, "-") {
-			parts := strings.SplitN(app.DestinationPort, "-", 2)
-			lo, _ := strconv.Atoi(parts[0])
-			hi, _ := strconv.Atoi(parts[1])
-			if dstPort < lo || dstPort > hi {
-				return false
-			}
-		} else {
-			p, _ := strconv.Atoi(app.DestinationPort)
-			if p != dstPort {
-				return false
-			}
-		}
-	}
-	return true
-}
-
-// handleTest dispatches test sub-commands (policy, routing, security-zone).
-func (c *CLI) handleTest(args []string) error {
-	if len(args) == 0 {
-		fmt.Println("test: specify a test command")
-		writeCompletionHelp(os.Stdout, treeHelpCandidates(operationalTree["test"].Children))
-		return nil
-	}
-
-	resolved, err := resolveCommand(args[0], keysFromTree(operationalTree["test"].Children))
-	if err != nil {
-		return err
-	}
-
-	switch resolved {
-	case "policy":
-		return c.testPolicy(args[1:])
-	case "routing":
-		return c.testRouting(args[1:])
-	case "security-zone":
-		return c.testSecurityZone(args[1:])
-	default:
-		return fmt.Errorf("unknown test command: %s", resolved)
-	}
-}
-
-// testPolicy performs a 5-tuple policy lookup similar to Junos "test policy".
-func (c *CLI) testPolicy(args []string) error {
-	cfg := c.store.ActiveConfig()
-	if cfg == nil {
-		fmt.Println("No active configuration")
-		return nil
-	}
-
-	var fromZone, toZone, srcIP, dstIP, proto string
-	var dstPort int
-	for i := 0; i < len(args); i++ {
-		switch args[i] {
-		case "from-zone":
-			if i+1 < len(args) {
-				i++
-				fromZone = args[i]
-			}
-		case "to-zone":
-			if i+1 < len(args) {
-				i++
-				toZone = args[i]
-			}
-		case "source-ip":
-			if i+1 < len(args) {
-				i++
-				srcIP = args[i]
-			}
-		case "destination-ip":
-			if i+1 < len(args) {
-				i++
-				dstIP = args[i]
-			}
-		case "destination-port":
-			if i+1 < len(args) {
-				i++
-				dstPort, _ = strconv.Atoi(args[i])
-			}
-		case "protocol":
-			if i+1 < len(args) {
-				i++
-				proto = args[i]
-			}
-		}
-	}
-
-	if fromZone == "" || toZone == "" {
-		fmt.Println("usage: test policy from-zone <zone> to-zone <zone>")
-		fmt.Println("       source-ip <ip> destination-ip <ip> destination-port <port> protocol <tcp|udp>")
-		return nil
-	}
-
-	parsedSrc := net.ParseIP(srcIP)
-	parsedDst := net.ParseIP(dstIP)
-
-	// Check zone-pair policies
-	for _, zpp := range cfg.Security.Policies {
-		if zpp.FromZone != fromZone || zpp.ToZone != toZone {
-			continue
-		}
-		for _, pol := range zpp.Policies {
-			if !matchPolicyAddr(pol.Match.SourceAddresses, parsedSrc, cfg) {
-				continue
-			}
-			if !matchPolicyAddr(pol.Match.DestinationAddresses, parsedDst, cfg) {
-				continue
-			}
-			if !matchPolicyApp(pol.Match.Applications, proto, dstPort, cfg) {
-				continue
-			}
-			action := "permit"
-			switch pol.Action {
-			case 1:
-				action = "deny"
-			case 2:
-				action = "reject"
-			}
-			fmt.Printf("Policy match:\n")
-			fmt.Printf("  From zone: %s\n  To zone:   %s\n", fromZone, toZone)
-			fmt.Printf("  Policy:    %s\n", pol.Name)
-			fmt.Printf("  Action:    %s\n", action)
-			if srcIP != "" {
-				fmt.Printf("  Source:    %s -> ", srcIP)
-			} else {
-				fmt.Printf("  Source:    any -> ")
-			}
-			if dstIP != "" {
-				fmt.Printf("%s", dstIP)
-			} else {
-				fmt.Printf("any")
-			}
-			if dstPort > 0 {
-				fmt.Printf(":%d", dstPort)
-			}
-			if proto != "" {
-				fmt.Printf(" [%s]", proto)
-			}
-			fmt.Println()
-			return nil
-		}
-	}
-
-	// Check global policies
-	for _, pol := range cfg.Security.GlobalPolicies {
-		if !matchPolicyAddr(pol.Match.SourceAddresses, parsedSrc, cfg) {
-			continue
-		}
-		if !matchPolicyAddr(pol.Match.DestinationAddresses, parsedDst, cfg) {
-			continue
-		}
-		if !matchPolicyApp(pol.Match.Applications, proto, dstPort, cfg) {
-			continue
-		}
-		action := "permit"
-		switch pol.Action {
-		case 1:
-			action = "deny"
-		case 2:
-			action = "reject"
-		}
-		fmt.Printf("Policy match (global):\n")
-		fmt.Printf("  Policy:    %s\n", pol.Name)
-		fmt.Printf("  Action:    %s\n", action)
-		return nil
-	}
-
-	fmt.Printf("Default deny (no matching policy for %s -> %s)\n", fromZone, toZone)
-	return nil
-}
-
-// testRouting looks up a destination in the routing table.
-func (c *CLI) testRouting(args []string) error {
-	if c.routing == nil {
-		fmt.Println("Routing manager not available")
-		return nil
-	}
-
-	var dest, instance string
-	for i := 0; i < len(args); i++ {
-		switch args[i] {
-		case "destination":
-			if i+1 < len(args) {
-				i++
-				dest = args[i]
-			}
-		case "instance":
-			if i+1 < len(args) {
-				i++
-				instance = args[i]
-			}
-		}
-	}
-
-	if dest == "" {
-		fmt.Println("usage: test routing destination <ip-or-prefix> [instance <name>]")
-		return nil
-	}
-
-	var entries []routing.RouteEntry
-	var err error
-	if instance != "" {
-		entries, err = c.routing.GetVRFRoutes(instance)
-	} else {
-		entries, err = c.routing.GetRoutes()
-	}
-	if err != nil {
-		return fmt.Errorf("get routes: %w", err)
-	}
-
-	// Normalize dest to CIDR for matching
-	filterCIDR := dest
-	if !strings.Contains(filterCIDR, "/") {
-		if strings.Contains(filterCIDR, ":") {
-			filterCIDR += "/128"
-		} else {
-			filterCIDR += "/32"
-		}
-	}
-	filterIP, _, filterErr := net.ParseCIDR(filterCIDR)
-	if filterErr != nil {
-		filterIP = net.ParseIP(dest)
-	}
-
-	// Find the best (longest prefix) match
-	var best *routing.RouteEntry
-	bestLen := -1
-	for i := range entries {
-		_, rNet, err := net.ParseCIDR(entries[i].Destination)
-		if err != nil {
-			continue
-		}
-		if filterIP != nil && rNet.Contains(filterIP) {
-			ones, _ := rNet.Mask.Size()
-			if ones > bestLen {
-				bestLen = ones
-				best = &entries[i]
-			}
-		}
-	}
-
-	if instance != "" {
-		fmt.Printf("Routing lookup in instance %s for %s:\n", instance, dest)
-	} else {
-		fmt.Printf("Routing lookup for %s:\n", dest)
-	}
-	if best == nil {
-		fmt.Println("  No matching route found")
-	} else {
-		fmt.Printf("  Destination: %s\n", best.Destination)
-		fmt.Printf("  Next-hop:    %s\n", best.NextHop)
-		fmt.Printf("  Interface:   %s\n", best.Interface)
-		fmt.Printf("  Protocol:    %s\n", best.Protocol)
-		fmt.Printf("  Preference:  %d\n", best.Preference)
-	}
-	return nil
-}
-
-// testSecurityZone looks up which zone an interface belongs to.
-func (c *CLI) testSecurityZone(args []string) error {
-	cfg := c.store.ActiveConfig()
-	if cfg == nil {
-		fmt.Println("No active configuration")
-		return nil
-	}
-
-	var ifName string
-	for i := 0; i < len(args); i++ {
-		if args[i] == "interface" && i+1 < len(args) {
-			i++
-			ifName = args[i]
-		}
-	}
-
-	if ifName == "" {
-		fmt.Println("usage: test security-zone interface <name>")
-		return nil
-	}
-
-	for zoneName, zone := range cfg.Security.Zones {
-		for _, iface := range zone.Interfaces {
-			if iface == ifName {
-				fmt.Printf("Interface %s belongs to zone: %s\n", ifName, zoneName)
-				if zone.Description != "" {
-					fmt.Printf("  Description: %s\n", zone.Description)
-				}
-				if zone.ScreenProfile != "" {
-					fmt.Printf("  Screen:      %s\n", zone.ScreenProfile)
-				}
-				if zone.HostInboundTraffic != nil {
-					if len(zone.HostInboundTraffic.SystemServices) > 0 {
-						fmt.Printf("  Host-inbound services: %s\n", strings.Join(zone.HostInboundTraffic.SystemServices, ", "))
-					}
-					if len(zone.HostInboundTraffic.Protocols) > 0 {
-						fmt.Printf("  Host-inbound protocols: %s\n", strings.Join(zone.HostInboundTraffic.Protocols, ", "))
-					}
-				}
-				return nil
-			}
-		}
-	}
-
-	fmt.Printf("Interface %s is not assigned to any security zone\n", ifName)
-	return nil
-}
-
-// handleMonitor dispatches monitor sub-commands.
-func (c *CLI) handleMonitor(args []string) error {
-	monTree := operationalTree["monitor"].Children
-	if len(args) == 0 {
-		fmt.Println("monitor:")
-		writeCompletionHelp(os.Stdout, treeHelpCandidates(monTree))
-		return nil
-	}
-
-	resolved, err := resolveCommand(args[0], keysFromTree(monTree))
-	if err != nil {
-		return err
-	}
-
-	switch resolved {
-	case "traffic":
-		return c.handleMonitorTraffic(args[1:])
-	case "interface":
-		return c.handleMonitorInterface(args[1:])
-	case "security":
-		return c.handleMonitorSecurity(args[1:])
-	default:
-		return fmt.Errorf("unknown monitor target: %s", resolved)
-	}
-}
-
-// handleMonitorTraffic wraps tcpdump for live packet capture.
-func (c *CLI) handleMonitorTraffic(args []string) error {
-	var iface, filter string
-	count := "0" // 0 = unlimited
-
-	for i := 0; i < len(args); i++ {
-		switch args[i] {
-		case "interface":
-			if i+1 < len(args) {
-				i++
-				iface = args[i]
-			}
-		case "matching":
-			if i+1 < len(args) {
-				i++
-				filter = args[i]
-			}
-		case "count":
-			if i+1 < len(args) {
-				i++
-				count = args[i]
-			}
-		}
-	}
-
-	if iface == "" {
-		fmt.Println("usage: monitor traffic interface <name> [matching <filter>] [count <N>]")
-		return nil
-	}
-
-	// Resolve fabric IPVLAN overlays to physical parent (#136).
-	origName := iface
-	iface = resolveFabricParent(iface)
-
-	// Warn about XDP redirect visibility on fabric interfaces (#138).
-	if strings.HasPrefix(origName, "fab") || strings.HasPrefix(origName, "em") {
-		fmt.Println("WARNING: XDP-redirected packets bypass AF_PACKET and will not appear in tcpdump.")
-		fmt.Println("For fabric redirect telemetry, use: show chassis cluster fabric statistics")
-		fmt.Println()
-	}
-
-	cmdArgs := []string{"tcpdump", "-i", iface, "-n", "-l"}
-	if count != "0" {
-		cmdArgs = append(cmdArgs, "-c", count)
-	}
-	if filter != "" {
-		cmdArgs = append(cmdArgs, filter)
-	}
-
-	fmt.Printf("Monitoring traffic on %s (Ctrl+C to stop)...\n", iface)
-
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
-	c.cmdMu.Lock()
-	c.cmdCancel = cancel
-	c.cmdMu.Unlock()
-	defer func() {
-		c.cmdMu.Lock()
-		c.cmdCancel = nil
-		c.cmdMu.Unlock()
-	}()
-
-	cmd := exec.CommandContext(ctx, cmdArgs[0], cmdArgs[1:]...)
-	cmd.Stdout = os.Stdout
-	cmd.Stderr = os.Stderr
-	err := cmd.Run()
-	if ctx.Err() != nil {
-		fmt.Println() // newline after ^C
-		return nil
-	}
-	return err
-}
-
-// showSystemUptime shows system uptime (like Junos "show system uptime").
-func fmtBytes(b uint64) string {
-	switch {
-	case b >= 1<<30:
-		return fmt.Sprintf("%.1fG", float64(b)/float64(1<<30))
-	case b >= 1<<20:
-		return fmt.Sprintf("%.1fM", float64(b)/float64(1<<20))
-	case b >= 1<<10:
-		return fmt.Sprintf("%.1fK", float64(b)/float64(1<<10))
-	default:
-		return fmt.Sprintf("%dB", b)
-	}
-}
-
-// showARP shows the kernel ARP table (like Junos "show arp no-resolve").
-func (c *CLI) handleShowIPv6(args []string) error {
-	if len(args) == 0 {
-		cmdtree.PrintTreeHelp("show ipv6:", operationalTree, "show", "ipv6")
-		return nil
-	}
-	switch args[0] {
-	case "neighbors":
-		return c.showIPv6Neighbors()
-	case "router-advertisement":
-		return c.showIPv6RouterAdvertisement()
-	default:
-		return fmt.Errorf("unknown show ipv6 target: %s", args[0])
-	}
-}
-
-// showIPv6Neighbors shows the kernel IPv6 neighbor cache (like Junos "show ipv6 neighbors").
-func neighState(state int) string {
-	switch state {
-	case netlink.NUD_REACHABLE:
-		return "reachable"
-	case netlink.NUD_STALE:
-		return "stale"
-	case netlink.NUD_DELAY:
-		return "delay"
-	case netlink.NUD_PROBE:
-		return "probe"
-	case netlink.NUD_FAILED:
-		return "failed"
-	case netlink.NUD_PERMANENT:
-		return "permanent"
-	case netlink.NUD_INCOMPLETE:
-		return "incomplete"
-	case netlink.NUD_NOARP:
-		return "noarp"
-	default:
-		return "unknown"
-	}
-}
-
-// showIPv6RouterAdvertisement shows RA configuration from the active config.
-
-func (c *CLI) buildInterfacesInput() cluster.InterfacesInput {
-	var input cluster.InterfacesInput
-	cfg := c.store.ActiveConfig()
-	if cfg == nil || cfg.Chassis.Cluster == nil {
-		return input
-	}
-	cc := cfg.Chassis.Cluster
-	input.ControlInterface = cc.ControlInterface
-	input.FabricInterface = cc.FabricInterface
-	if fabIfc, ok := cfg.Interfaces.Interfaces[cc.FabricInterface]; ok {
-		input.FabricMembers = fabIfc.FabricMembers
-	}
-	input.Fabric1Interface = cc.Fabric1Interface
-	if fab1Ifc, ok := cfg.Interfaces.Interfaces[cc.Fabric1Interface]; ok {
-		input.Fabric1Members = fab1Ifc.FabricMembers
-	}
-
-	// Build RETH info from config.
-	rethMap := cfg.RethToPhysical() // reth-name -> physical-member
-	for name, ifc := range cfg.Interfaces.Interfaces {
-		if ifc.RedundancyGroup > 0 && strings.HasPrefix(name, "reth") {
-			status := "Up"
-			// Check physical member link state.
-			if phys, ok := rethMap[name]; ok {
-				linuxName := config.LinuxIfName(phys)
-				link, err := netlink.LinkByName(linuxName)
-				if err != nil || (link.Attrs().OperState != netlink.OperUp &&
-					link.Attrs().Flags&net.FlagUp == 0) {
-					status = "Down"
-				}
-			}
-			input.Reths = append(input.Reths, cluster.RethInfo{
-				Name:            name,
-				RedundancyGroup: ifc.RedundancyGroup,
-				Status:          status,
-			})
-		}
-	}
-	sort.Slice(input.Reths, func(i, j int) bool { return input.Reths[i].Name < input.Reths[j].Name })
-
-	// Build local interface monitor info.
-	localMonMap := make(map[string]bool) // track which interfaces are local
-	monStatuses := make(map[int][]routing.InterfaceMonitorStatus)
-	if c.routing != nil {
-		if ms := c.routing.InterfaceMonitorStatuses(); ms != nil {
-			monStatuses = ms
-		}
-	}
-	for _, rg := range cc.RedundancyGroups {
-		if statuses, ok := monStatuses[rg.ID]; ok {
-			for _, st := range statuses {
-				input.Monitors = append(input.Monitors, cluster.InterfaceMonitorInfo{
-					Interface:       st.Interface,
-					Weight:          st.Weight,
-					Up:              st.Up,
-					RedundancyGroup: rg.ID,
-				})
-				localMonMap[st.Interface] = true
-			}
-		} else {
-			for _, mon := range rg.InterfaceMonitors {
-				input.Monitors = append(input.Monitors, cluster.InterfaceMonitorInfo{
-					Interface:       mon.Interface,
-					Weight:          mon.Weight,
-					Up:              true, // unknown, assume up
-					RedundancyGroup: rg.ID,
-				})
-				localMonMap[mon.Interface] = true
-			}
-		}
-	}
-
-	// Build peer interface monitor info from heartbeat.
-	if c.cluster != nil {
-		peerLive := c.cluster.PeerMonitorStatuses()
-		peerMap := make(map[string]bool)
-		for _, pm := range peerLive {
-			peerMap[pm.Interface] = true
-			input.PeerMonitors = append(input.PeerMonitors, pm)
-		}
-		// Fill config-only peer monitors (not local, not in heartbeat) as down.
-		for _, rg := range cc.RedundancyGroups {
-			for _, mon := range rg.InterfaceMonitors {
-				if localMonMap[mon.Interface] {
-					continue
-				}
-				if peerMap[mon.Interface] {
-					continue
-				}
-				input.PeerMonitors = append(input.PeerMonitors, cluster.InterfaceMonitorInfo{
-					Interface:       mon.Interface,
-					Weight:          mon.Weight,
-					Up:              false,
-					RedundancyGroup: rg.ID,
-				})
-			}
-		}
-	}
-
-	return input
-}
-
-func (c *CLI) userspaceDataplaneStatus() (dpuserspace.ProcessStatus, error) {
-	provider, ok := c.dp.(interface {
-		Status() (dpuserspace.ProcessStatus, error)
-	})
-	if !ok {
-		return dpuserspace.ProcessStatus{}, fmt.Errorf("userspace status unavailable")
-	}
-	return provider.Status()
-}
-
-func (c *CLI) userspaceDataplaneControl() (interface {
-	Status() (dpuserspace.ProcessStatus, error)
-	SetForwardingArmed(bool) (dpuserspace.ProcessStatus, error)
-	SetQueueState(uint32, bool, bool) (dpuserspace.ProcessStatus, error)
-	SetBindingState(uint32, bool, bool) (dpuserspace.ProcessStatus, error)
-	InjectPacket(dpuserspace.InjectPacketRequest) (dpuserspace.ProcessStatus, error)
-}, error) {
-	provider, ok := c.dp.(interface {
-		Status() (dpuserspace.ProcessStatus, error)
-		SetForwardingArmed(bool) (dpuserspace.ProcessStatus, error)
-		SetQueueState(uint32, bool, bool) (dpuserspace.ProcessStatus, error)
-		SetBindingState(uint32, bool, bool) (dpuserspace.ProcessStatus, error)
-		InjectPacket(dpuserspace.InjectPacketRequest) (dpuserspace.ProcessStatus, error)
-	})
-	if !ok {
-		return nil, fmt.Errorf("userspace dataplane control unavailable")
-	}
-	return provider, nil
-}
-
-func fmtPref(p int) string {
-	if p == 0 {
-		return "-"
-	}
-	return strconv.Itoa(p)
-}
-
-// showForwardingOptions displays forwarding/sampling configuration.
-func (c *CLI) handleRequest(args []string) error {
-	if len(args) == 0 {
-		fmt.Println("request:")
-		writeCompletionHelp(os.Stdout, treeHelpCandidates(operationalTree["request"].Children))
-		return nil
-	}
-
-	switch args[0] {
-	case "chassis":
-		return c.handleRequestChassis(args[1:])
-	case "dhcp":
-		return c.handleRequestDHCP(args[1:])
-	case "protocols":
-		return c.handleRequestProtocols(args[1:])
-	case "security":
-		return c.handleRequestSecurity(args[1:])
-	case "system":
-		return c.handleRequestSystem(args[1:])
-	default:
-		return fmt.Errorf("unknown request target: %s", args[0])
-	}
-}
-
-func (c *CLI) handleRequestChassis(args []string) error {
-	if len(args) == 0 || args[0] != "cluster" {
-		fmt.Println("request chassis:")
-		writeCompletionHelp(os.Stdout, treeHelpCandidates(operationalTree["request"].Children["chassis"].Children))
-		return nil
-	}
-	args = args[1:] // consume "cluster"
-	if len(args) == 0 {
-		fmt.Println("request chassis cluster:")
-		writeCompletionHelp(os.Stdout, treeHelpCandidates(operationalTree["request"].Children["chassis"].Children["cluster"].Children))
-		return nil
-	}
-	switch args[0] {
-	case "failover":
-		return c.handleRequestChassisClusterFailover(args[1:])
-	case "data-plane":
-		return c.handleRequestChassisClusterDataPlane(args[1:])
-	default:
-		fmt.Println("request chassis cluster:")
-		writeCompletionHelp(os.Stdout, treeHelpCandidates(operationalTree["request"].Children["chassis"].Children["cluster"].Children))
-		return nil
-	}
-}
-
-func (c *CLI) handleRequestChassisClusterFailover(args []string) error {
-	if c.cluster == nil {
-		return fmt.Errorf("cluster not configured")
-	}
-	// "request chassis cluster failover reset redundancy-group <N>"
-	if len(args) >= 1 && args[0] == "reset" {
-		if len(args) < 3 || args[1] != "redundancy-group" {
-			return fmt.Errorf("usage: request chassis cluster failover reset redundancy-group <N>")
-		}
-		rgID, err := strconv.Atoi(args[2])
-		if err != nil {
-			return fmt.Errorf("invalid redundancy-group ID: %s", args[2])
-		}
-		if err := c.cluster.ResetFailover(rgID); err != nil {
-			return err
-		}
-		fmt.Printf("Failover reset for redundancy group %d\n", rgID)
-		return nil
-	}
-
-	// "request chassis cluster failover data node <N>"
-	if len(args) >= 3 && args[0] == "data" && args[1] == "node" {
-		targetNode, err := strconv.Atoi(args[2])
-		if err != nil {
-			return fmt.Errorf("invalid node ID: %s", args[2])
-		}
-		if !cluster.IsSupportedClusterNodeID(targetNode) {
-			return fmt.Errorf("unsupported cluster failover target node %d", targetNode)
-		}
-		localNode := c.cluster.NodeID()
-		if targetNode != localNode {
-			message, err := c.requestPeerSystemAction(
-				context.Background(),
-				fmt.Sprintf("cluster-failover-data:node%d", targetNode),
-			)
-			if err != nil {
-				return err
-			}
-			fmt.Println(message)
-			return nil
-		}
-
-		dataRGs := c.cluster.DataGroupIDs()
-		if len(dataRGs) == 0 {
-			return fmt.Errorf("no data redundancy groups configured")
-		}
-		moveRGs := make([]int, 0, len(dataRGs))
-		for _, rgID := range dataRGs {
-			if !c.cluster.IsLocalPrimary(rgID) {
-				moveRGs = append(moveRGs, rgID)
-			}
-		}
-		if len(moveRGs) == 0 {
-			fmt.Printf("All data redundancy groups are already primary on node %d\n", targetNode)
-			return nil
-		}
-		if len(moveRGs) == 1 {
-			if err := c.cluster.RequestPeerFailover(moveRGs[0]); err != nil {
-				return err
-			}
-		} else {
-			if err := c.cluster.RequestPeerFailoverBatch(moveRGs); err != nil {
-				return err
-			}
-		}
-		fmt.Printf("Manual failover completed for data redundancy groups %v (transfer committed)\n", moveRGs)
-		return nil
-	}
-
-	// "request chassis cluster failover redundancy-group <N> [node <N>]"
-	if len(args) >= 2 && args[0] == "redundancy-group" {
-		rgID, err := strconv.Atoi(args[1])
-		if err != nil {
-			return fmt.Errorf("invalid redundancy-group ID: %s", args[1])
-		}
-
-		// If "node <N>" is specified, route to the correct node.
-		if len(args) >= 4 && args[2] == "node" {
-			targetNode, err := strconv.Atoi(args[3])
-			if err != nil {
-				return fmt.Errorf("invalid node ID: %s", args[3])
-			}
-			localNode := c.cluster.NodeID()
-			if targetNode == localNode {
-				if err := c.cluster.RequestPeerFailover(rgID); err != nil {
-					return err
-				}
-				fmt.Printf("Manual failover completed for redundancy group %d (transfer committed)\n", rgID)
-				return nil
-			}
-			message, err := c.requestPeerSystemAction(
-				context.Background(),
-				fmt.Sprintf("cluster-failover:%d:node%d", rgID, targetNode),
-			)
-			if err != nil {
-				return err
-			}
-			fmt.Println(message)
-			return nil
-		}
-
-		if err := c.cluster.ManualFailover(rgID); err != nil {
-			return err
-		}
-		fmt.Printf("Manual failover triggered for redundancy group %d\n", rgID)
-		return nil
-	}
-
-	return fmt.Errorf("usage: request chassis cluster failover {redundancy-group <N> [node <N>] | data node <N>}")
-}
-
-func (c *CLI) handleRequestChassisClusterDataPlane(args []string) error {
-	if len(args) == 0 || args[0] != "userspace" {
-		fmt.Println("request chassis cluster data-plane:")
-		writeCompletionHelp(os.Stdout, treeHelpCandidates(operationalTree["request"].Children["chassis"].Children["cluster"].Children["data-plane"].Children))
-		return nil
-	}
-	provider, err := c.userspaceDataplaneControl()
-	if err != nil {
-		return err
-	}
-	args = args[1:]
-
-	var status dpuserspace.ProcessStatus
-	switch {
-	case len(args) > 0 && args[0] == "inject-packet":
-		slot, mode, extra, err := dpuserspace.ParseInjectPacketCommand(args)
-		if err != nil {
-			return err
-		}
-		status, err = provider.Status()
-		if err != nil {
-			return err
-		}
-		req, err := dpuserspace.BuildInjectPacketRequest(slot, mode, extra, status)
-		if err != nil {
-			return err
-		}
-		status, err = provider.InjectPacket(req)
-		if err != nil {
-			return err
-		}
-	case len(args) > 0 && args[0] == "forwarding":
-		armed, err := dpuserspace.ParseForwardingCommand(args)
-		if err != nil {
-			return err
-		}
-		status, err = provider.SetForwardingArmed(armed)
-		if err != nil {
-			return err
-		}
-	case len(args) > 0 && args[0] == "queue":
-		queueID, registered, armed, err := dpuserspace.ParseQueueCommand(args)
-		if err != nil {
-			return err
-		}
-		status, err = provider.SetQueueState(queueID, registered, armed)
-		if err != nil {
-			return err
-		}
-	case len(args) > 0 && args[0] == "binding":
-		slot, registered, armed, err := dpuserspace.ParseBindingCommand(args)
-		if err != nil {
-			return err
-		}
-		status, err = provider.SetBindingState(slot, registered, armed)
-		if err != nil {
-			return err
-		}
-	default:
-		fmt.Println("request chassis cluster data-plane userspace:")
-		writeCompletionHelp(os.Stdout, treeHelpCandidates(operationalTree["request"].Children["chassis"].Children["cluster"].Children["data-plane"].Children["userspace"].Children))
-		return nil
-	}
-	fmt.Print(dpuserspace.FormatStatusSummary(status))
-	fmt.Println()
-	fmt.Print(dpuserspace.FormatBindings(status))
-	return nil
-}
-
-func (c *CLI) handleRequestDHCP(args []string) error {
-	if len(args) == 0 || args[0] != "renew" {
-		fmt.Println("request dhcp:")
-		writeCompletionHelp(os.Stdout, treeHelpCandidates(operationalTree["request"].Children["dhcp"].Children))
-		return nil
-	}
-	if len(args) < 2 {
-		return fmt.Errorf("usage: request dhcp renew <interface>")
-	}
-	if c.dhcp == nil {
-		return fmt.Errorf("DHCP manager not available")
-	}
-	if err := c.dhcp.Renew(args[1]); err != nil {
-		return err
-	}
-	fmt.Printf("DHCP renewal initiated on %s\n", args[1])
-	return nil
-}
-
-func (c *CLI) handleRequestProtocols(args []string) error {
-	if len(args) == 0 {
-		fmt.Println("request protocols:")
-		writeCompletionHelp(os.Stdout, treeHelpCandidates(operationalTree["request"].Children["protocols"].Children))
-		return nil
-	}
-	if c.frr == nil {
-		return fmt.Errorf("FRR manager not available")
-	}
-	switch args[0] {
-	case "ospf":
-		if len(args) < 2 || args[1] != "clear" {
-			fmt.Println("request protocols ospf:")
-			writeCompletionHelp(os.Stdout, treeHelpCandidates(operationalTree["request"].Children["protocols"].Children["ospf"].Children))
-			return nil
-		}
-		output, err := c.frr.ExecVtysh("clear ip ospf process")
-		if err != nil {
-			return fmt.Errorf("clear OSPF: %w", err)
-		}
-		if output != "" {
-			fmt.Print(output)
-		}
-		fmt.Println("OSPF process cleared")
-		return nil
-	case "bgp":
-		if len(args) < 2 || args[1] != "clear" {
-			fmt.Println("request protocols bgp:")
-			writeCompletionHelp(os.Stdout, treeHelpCandidates(operationalTree["request"].Children["protocols"].Children["bgp"].Children))
-			return nil
-		}
-		output, err := c.frr.ExecVtysh("clear bgp * soft")
-		if err != nil {
-			return fmt.Errorf("clear BGP: %w", err)
-		}
-		if output != "" {
-			fmt.Print(output)
-		}
-		fmt.Println("BGP sessions cleared (soft reset)")
-		return nil
-	default:
-		return fmt.Errorf("unknown request protocols target: %s", args[0])
-	}
-}
-
-func (c *CLI) handleRequestSystem(args []string) error {
-	if len(args) == 0 {
-		fmt.Println("request system:")
-		writeCompletionHelp(os.Stdout, treeHelpCandidates(operationalTree["request"].Children["system"].Children))
-		return nil
-	}
-
-	switch args[0] {
-	case "reboot":
-		fmt.Print("Reboot the system? [yes,no] (no) ")
-		c.rl.SetPrompt("")
-		line, err := c.rl.Readline()
-		c.rl.SetPrompt(c.operationalPrompt())
-		if err != nil || strings.TrimSpace(strings.ToLower(line)) != "yes" {
-			fmt.Println("Reboot cancelled")
-			return nil
-		}
-		fmt.Println("System going down for reboot NOW!")
-		cmd := exec.Command("systemctl", "reboot")
-		return cmd.Run()
-
-	case "halt":
-		fmt.Print("Halt the system? [yes,no] (no) ")
-		c.rl.SetPrompt("")
-		line, err := c.rl.Readline()
-		c.rl.SetPrompt(c.operationalPrompt())
-		if err != nil || strings.TrimSpace(strings.ToLower(line)) != "yes" {
-			fmt.Println("Halt cancelled")
-			return nil
-		}
-		fmt.Println("System halting NOW!")
-		cmd := exec.Command("systemctl", "halt")
-		return cmd.Run()
-
-	case "power-off":
-		fmt.Print("Power off the system? [yes,no] (no) ")
-		c.rl.SetPrompt("")
-		line, err := c.rl.Readline()
-		c.rl.SetPrompt(c.operationalPrompt())
-		if err != nil || strings.TrimSpace(strings.ToLower(line)) != "yes" {
-			fmt.Println("Power-off cancelled")
-			return nil
-		}
-		fmt.Println("System powering off NOW!")
-		cmd := exec.Command("systemctl", "poweroff")
-		return cmd.Run()
-
-	case "zeroize":
-		fmt.Println("WARNING: This will erase all configuration and return to factory defaults.")
-		fmt.Print("Zeroize the system? [yes,no] (no) ")
-		c.rl.SetPrompt("")
-		line, err := c.rl.Readline()
-		c.rl.SetPrompt(c.operationalPrompt())
-		if err != nil || strings.TrimSpace(strings.ToLower(line)) != "yes" {
-			fmt.Println("Zeroize cancelled")
-			return nil
-		}
-
-		// Remove active and candidate configs, rollback history
-		configDir := "/etc/bpfrx"
-		files, _ := os.ReadDir(configDir)
-		for _, f := range files {
-			if strings.HasSuffix(f.Name(), ".conf") || strings.HasPrefix(f.Name(), "rollback") {
-				os.Remove(configDir + "/" + f.Name())
-			}
-		}
-
-		// Remove BPF pins
-		os.RemoveAll("/sys/fs/bpf/bpfrx")
-
-		// Remove managed networkd files
-		ndFiles, _ := os.ReadDir("/etc/systemd/network")
-		for _, f := range ndFiles {
-			if strings.HasPrefix(f.Name(), "10-bpfrx-") {
-				os.Remove("/etc/systemd/network/" + f.Name())
-			}
-		}
-
-		// Remove FRR managed section
-		exec.Command("systemctl", "stop", "bpfrxd").Run()
-
-		fmt.Println("System zeroized. Configuration erased.")
-		fmt.Println("Reboot to complete factory reset.")
-		return nil
-
-	case "configuration":
-		return c.handleRequestSystemConfiguration(args[1:])
-
-	case "software":
-		return c.handleRequestSystemSoftware(args[1:])
-
-	default:
-		return fmt.Errorf("unknown request system command: %s", args[0])
-	}
-}
-
-func (c *CLI) handleRequestSystemSoftware(args []string) error {
-	if len(args) == 0 {
-		fmt.Println("request system software:")
-		writeCompletionHelp(os.Stdout, treeHelpCandidates(operationalTree["request"].Children["system"].Children["software"].Children))
-		return nil
-	}
-
-	if args[0] != "in-service-upgrade" {
-		return fmt.Errorf("unknown request system software command: %s", args[0])
-	}
-
-	if c.cluster == nil {
-		fmt.Println("Cluster not configured")
-		return nil
-	}
-
-	fmt.Println("WARNING: This will force this node to secondary for all redundancy groups.")
-	fmt.Print("Proceed with in-service upgrade? [yes,no] (no) ")
-	c.rl.SetPrompt("")
-	line, err := c.rl.Readline()
-	c.rl.SetPrompt(c.operationalPrompt())
-	if err != nil || strings.TrimSpace(strings.ToLower(line)) != "yes" {
-		fmt.Println("ISSU cancelled")
-		return nil
-	}
-
-	if err := c.cluster.ForceSecondary(); err != nil {
-		return fmt.Errorf("ISSU: %v", err)
-	}
-
-	fmt.Println("Node is now secondary for all redundancy groups.")
-	fmt.Println("Traffic has been drained to peer.")
-	fmt.Println("You may now replace the binary and restart the service:")
-	fmt.Println("  systemctl stop bpfrxd && <replace binary> && systemctl start bpfrxd")
-	return nil
-}
-
-func (c *CLI) handleRequestSystemConfiguration(args []string) error {
-	if len(args) == 0 {
-		fmt.Println("request system configuration:")
-		writeCompletionHelp(os.Stdout, treeHelpCandidates(operationalTree["request"].Children["system"].Children["configuration"].Children))
-		return nil
-	}
-
-	if args[0] != "rescue" {
-		return fmt.Errorf("unknown request system configuration command: %s", args[0])
-	}
-
-	if len(args) < 2 {
-		fmt.Println("request system configuration rescue:")
-		writeCompletionHelp(os.Stdout, treeHelpCandidates(operationalTree["request"].Children["system"].Children["configuration"].Children["rescue"].Children))
-		return nil
-	}
-
-	switch args[1] {
-	case "save":
-		if err := c.store.SaveRescueConfig(); err != nil {
-			return err
-		}
-		fmt.Println("Rescue configuration saved")
-		return nil
-
-	case "delete":
-		if err := c.store.DeleteRescueConfig(); err != nil {
-			return err
-		}
-		fmt.Println("Rescue configuration deleted")
-		return nil
-
-	default:
-		return fmt.Errorf("unknown request system configuration rescue command: %s", args[1])
-	}
-}
-
-func (c *CLI) handleRequestSecurity(args []string) error {
-	if len(args) == 0 {
-		fmt.Println("request security:")
-		writeCompletionHelp(os.Stdout, treeHelpCandidates(operationalTree["request"].Children["security"].Children))
-		return nil
-	}
-	if args[0] != "ipsec" {
-		return fmt.Errorf("unknown request security target: %s", args[0])
-	}
-	if len(args) < 3 || args[1] != "sa" || args[2] != "clear" {
-		fmt.Println("request security ipsec sa:")
-		writeCompletionHelp(os.Stdout, treeHelpCandidates(operationalTree["request"].Children["security"].Children["ipsec"].Children["sa"].Children))
-		return nil
-	}
-	if c.ipsec == nil {
-		return fmt.Errorf("IPsec manager not available")
-	}
-	count, err := c.ipsec.TerminateAllSAs()
-	if err != nil {
-		return err
-	}
-	fmt.Printf("Cleared %d IPsec SA(s)\n", count)
-	return nil
 }
